Cleaned field from script subsystem and grouped into Field catalogue
