+++ FF8/Field/init_debugger_fields.cs
